Refactor folder structure to be able to include argument specific utils
