+++ commands/slixer/slixer_test.py
@@ -1,33 +1,37 @@
 import pytest
 import argparse
 from unittest.mock import patch
-from commands.slixer.command import slixer
+from commands.slixer.slixer import slixer
 
 
 @pytest.fixture
 def mock_audio_file_path_validate():
-    with patch("arguments.audio_file_path.validate") as mock:
+    with patch("arguments.audio_file_path.audio_file_path.validate") as mock:
         mock.return_value = True
         yield mock
 
 
 @pytest.fixture
 def mock_timestamps_file_path_validate():
-    with patch("arguments.timestamps_file_path.validate") as mock:
+    with patch(
+        "arguments.timestamps_file_path.timestamps_file_path.validate"
+    ) as mock:
         mock.return_value = True
         yield mock
 
 
 @pytest.fixture
 def mock_timestamps_file_path_extract_timestamps():
-    with patch("arguments.timestamps_file_path.extract_timestamps") as mock:
+    with patch(
+        "arguments.timestamps_file_path.timestamps_file_path.extract_timestamps"  # noqa: E501
+    ) as mock:
         mock.return_value = []
         yield mock
 
 
 @pytest.fixture
 def mock_slixer_utils_split_audio_file():
-    with patch("commands.slixer.command.split_audio_file") as mock:
+    with patch("commands.slixer.slixer.split_audio_file") as mock:
         yield mock
